requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
openai==1.35.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
yarl==1.9.4